split up meta into separate files
