+++ lib/clusters.dart
@@ -1,3 +1,3 @@
 library turf_clusters;
 
-export 'package:turf/src/clusters.dart'; //TODO: should we show helpers too?
+export 'src/meta/cluster.dart';
