+++ lib/meta.dart
@@ -1,4 +1,9 @@
 library turf_meta;
 
-export 'src/meta.dart';
+export 'src/meta/cluster.dart';
+export 'src/meta/coord.dart';
+export 'src/meta/feature.dart';
+export 'src/meta/flatten.dart';
+export 'src/meta/geom.dart';
 export 'src/line_segment.dart' show segmentEach, segmentReduce;
+export 'src/meta/prop.dart';
