+++ lib/src/meta/short_circuit.dart
@@ -1,0 +1,5 @@
+/// A simple class to manage short circuiting from *Each functions while still
+/// allowing an Exception to be thrown and raised
+class ShortCircuit {
+  ShortCircuit();
+}
